restructure project, add placeholder components for future features
restructure project, add placeholder components for future features
