+++ frontend/src/app/app.component.html
@@ -1,4 +1,3 @@
 <div class="container" style="width: 600px">
   <app-navbar></app-navbar>
-  <router-outlet></router-outlet>
 </div>
