+++ frontend/src/app/app.module.ts
@@ -1,13 +1,13 @@
 import {NgModule} from "@angular/core";
 import {AppComponent} from "./app.component";
-import {FileUploadComponent} from "./components/file-upload/file-upload.component";
+import {FileUploadComponent} from "./features/components/file-upload/file-upload.component";
 import {BrowserModule} from "@angular/platform-browser";
 import {HttpClientModule} from "@angular/common/http";
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
-import {ImageEntryComponent} from "./components/file-upload/image-entry/image-entry.component";
+import {ImageEntryComponent} from "./features/components/file-upload/image-entry/image-entry.component";
 import {NgOptimizedImage} from "@angular/common";
-import {NavbarComponent} from "./components/navbar/navbar.component";
-import {RouterModule, RouterOutlet, Routes} from "@angular/router";
+import {NavbarComponent} from "./core/components/navbar/navbar.component";
+import {RouterModule} from "@angular/router";
 import {routes} from "./app.routes";
 
 @NgModule({
