+++ frontend/src/app/app.routes.ts
@@ -1,9 +1,15 @@
 import { Routes } from '@angular/router';
-import {AppComponent} from "./app.component";
-import {SchedulerComponent} from "./components/scheduler/scheduler.component";
-import {FileUploadComponent} from "./components/file-upload/file-upload.component";
+import {SchedulerComponent} from "./features/components/scheduler/scheduler.component";
+import {FileUploadComponent} from "./features/components/file-upload/file-upload.component";
+import {HomePageComponent} from "./core/components/home-page/home-page.component";
+import {PageNotFoundComponent} from "./core/components/page-not-found/page-not-found.component";
 
 export const routes: Routes = [
+  {
+    path: '',
+    component: HomePageComponent,
+    title: 'Home'
+  },
   {
     path: 'images',
     component: FileUploadComponent,
@@ -13,5 +19,10 @@ export const routes: Routes = [
     path: 'scheduler',
     component: SchedulerComponent,
     title: 'Scheduler'
+  },
+  {
+    path: '**',
+    component: PageNotFoundComponent,
+    title: 'Page not found!'
   },
 ];
