+++ frontend/src/app/core/components/home-page/home-page.component.html
@@ -1,0 +1,1 @@
+<p>home-page works!</p>
