+++ frontend/src/app/core/components/navbar/navbar.component.html
@@ -29,3 +29,5 @@
     </div>
   </div>
 </nav>
+
+<router-outlet></router-outlet>
