+++ frontend/src/app/core/components/navbar/navbar.component.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
-import {RouterLink} from "@angular/router";
+import {RouterLink, RouterOutlet} from "@angular/router";
 
 @Component({
   selector: 'app-navbar',
   standalone: true,
   imports: [
     RouterLink,
+    RouterOutlet,
   ],
   templateUrl: './navbar.component.html',
   styleUrl: './navbar.component.css'
