+++ frontend/src/app/core/components/page-not-found/page-not-found.component.html
@@ -1,0 +1,1 @@
+<p>page-not-found works!</p>
