+++ frontend/src/app/features/components/file-upload/image-entry/image-entry.component.ts
