+++ frontend/src/app/features/components/scheduler/scheduler.component.ts
