+++ frontend/src/app/features/models/image.model.ts
