+++ frontend/src/app/features/services/image-upload.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import {HttpClient, HttpEvent, HttpHeaders, HttpRequest} from "@angular/common/http";
 import {Observable} from "rxjs";
-import { environment } from '../../environment/environment';
+import { environment } from '../../../environment/environment';
 
 @Injectable({
   providedIn: 'root'
